+++ tests/ApiTest.php
@@ -6,12 +6,14 @@ namespace Ely\Mojang\Test;
 use Ely\Mojang\Api;
 use Ely\Mojang\Exception\ForbiddenException;
 use Ely\Mojang\Exception\NoContentException;
+use Ely\Mojang\Exception\OperationException;
 use Ely\Mojang\Middleware\ResponseConverterMiddleware;
 use Ely\Mojang\Middleware\RetryMiddleware;
 use Ely\Mojang\Response\ApiStatus;
 use Ely\Mojang\Response\NameHistoryItem;
 use Ely\Mojang\Response\ProfileInfo;
 use Ely\Mojang\Response\Properties\TexturesProperty;
+use Ely\Mojang\Response\QuestionResponse;
 use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Handler\MockHandler;
@@ -309,7 +311,7 @@ class ApiTest extends TestCase {
         }
 
         $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage('You cannot request more than 100 names per request');
+        $this->expectExceptionMessage('You cannot request more than 10 names per request');
         $this->api->playernamesToUuids($names);
     }
 
@@ -607,6 +609,130 @@ class ApiTest extends TestCase {
             }
         };
         $this->assertInstanceOf(ClientInterface::class, $child->getDefaultClient());
+    }
+
+    public function testIsSecurityQuestionsNeeded() {
+        $this->mockHandler->append(new Response(204));
+        $this->expectException(NoContentException::class);
+        $this->api->isSecurityQuestionsNeeded('mocked access token');
+    }
+
+    public function testIsSecurityQuestionsNeededOperationException() {
+        $this->mockHandler->append($this->createResponse(200, [
+            'error' => 'ForbiddenOperationException',
+            'errorMessage' => 'Current IP is not secured',
+        ]));
+        $this->expectException(OperationException::class);
+        $this->api->isSecurityQuestionsNeeded('mocked access token');
+    }
+
+    public function testQuestions() {
+        $this->mockHandler->append($this->createResponse(200, [
+            [
+                'answer' => [
+                    'id' => 123,
+                ],
+                'question' => [
+                    'id' => 1,
+                    'question' => 'What is your favorite pet\'s name?',
+                ],
+            ],
+            [
+                'answer' => [
+                    'id' => 456,
+                ],
+                'question' => [
+                    'id' => 2,
+                    'question' => 'What is your favorite movie?',
+                ],
+            ],
+            [
+                'answer' => [
+                    'id' => 789,
+                ],
+                'question' => [
+                    'id' => 3,
+                    'question' => 'What is your favorite author\'s last name?',
+                ],
+            ],
+        ]));
+        $result = $this->api->questions('mocked access token');
+        /** @var \Psr\Http\Message\RequestInterface $request */
+        $request = $this->history[0]['request'];
+        $this->assertSame('Bearer mocked access token', $request->getHeaderLine('Authorization'));
+
+        foreach ($result as $question) {
+            $this->assertInstanceOf(QuestionResponse::class, $question);
+        }
+
+        /** @var QuestionResponse $firstQuestion */
+        $firstQuestion = $result[0];
+        $this->assertSame(123, $firstQuestion->getAnswerId());
+        $this->assertSame(1, $firstQuestion->getQuestionId());
+        $this->assertSame('What is your favorite pet\'s name?', $firstQuestion->getQuestion());
+
+        /** @var QuestionResponse $secondQuestion */
+        $secondQuestion = $result[1];
+        $this->assertSame(456, $secondQuestion->getAnswerId());
+        $this->assertSame(2, $secondQuestion->getQuestionId());
+        $this->assertSame('What is your favorite movie?', $secondQuestion->getQuestion());
+
+        /** @var QuestionResponse $thirdQuestion */
+        $thirdQuestion = $result[2];
+        $this->assertSame(789, $thirdQuestion->getAnswerId());
+        $this->assertSame(3, $thirdQuestion->getQuestionId());
+        $this->assertSame('What is your favorite author\'s last name?', $thirdQuestion->getQuestion());
+    }
+
+    public function testAnswerOperationException() {
+        $this->mockHandler->append($this->createResponse(200, [
+            'error' => 'ForbiddenOperationException',
+            'errorMessage' => 'At least one answer was incorrect',
+        ]));
+        $result = $this->api->answer('mocked access token', [
+            [
+                'id' => 123,
+                'answer' => 'foo',
+            ],
+            [
+                'id' => 456,
+                'answer' => 'bar',
+            ],
+        ]);
+        $this->assertFalse($result);
+    }
+
+    public function testAnswer() {
+        $this->mockHandler->append(new Response(204));
+        $this->expectException(NoContentException::class);
+        $this->api->isSecurityQuestionsNeeded('mocked access token');
+    }
+
+    public function testStatistics() {
+        $this->mockHandler->append($this->createResponse(200, [
+            'total' => 145,
+            'last24h' => 13,
+            'saleVelocityPerSeconds' => 1.32,
+        ]));
+        $result = $this->api->statistics([
+            'item_sold_minecraft',
+            'prepaid_card_redeemed_minecraft',
+            'item_sold_cobalt',
+            'item_sold_scrolls',
+        ]);
+        /** @var \Psr\Http\Message\RequestInterface $request */
+        $request = $this->history[0]['request'];
+        $params = json_decode($request->getBody()->getContents(), true);
+        $this->assertSame([
+            'item_sold_minecraft',
+            'prepaid_card_redeemed_minecraft',
+            'item_sold_cobalt',
+            'item_sold_scrolls',
+        ], $params['metricKeys']);
+
+        $this->assertSame(145, $result->getTotal());
+        $this->assertSame(13, $result->getLast24H());
+        $this->assertSame(1.32, $result->getSaleVelocityPerSeconds());
     }
 
     private function createResponse(int $statusCode, array $response): ResponseInterface {